Fix keyboard state recovery
The recovered state for each letter was the state of the most
recently-encountered tile of that letter. However, that's not always
correct.

E.g. for a guess 'PARRY' for the answer 'PARSE', the first R would be
green and the second grey; however this algorithm would colour the key
grey.

This is corrected by only recolouring a key during the recovery if the
incoming state has a higher priority than the existing one.